add hook customizado
