+++ src/components/Produtos/index.jsx
@@ -1,31 +1,9 @@
 import Titulo from "@/components/Titulo";
-import { CarrinhoContext } from "@/context/CarrinhoContext";
 import produtos from "@/mocks/produtos.json";
-import React, { useContext } from "react";
+import React from "react";
 import Produto from "./Produto";
 
 const Produtos = () => {
-  const { carrinho, setCarrinho } = useContext(CarrinhoContext);
-  function adicionarProduto(novoProduto) {
-    const temProduto = carrinho.some((itemDoCarrinho) => {
-      itemDoCarrinho.id === novoProduto.id;
-    });
-
-    if (!temProduto) {
-      novoProduto.quantidade = 1;
-      return setCarrinho((carrinhoAnterior) => [
-        ...carrinhoAnterior,
-        novoProduto,
-      ]);
-    }
-    setCarrinho((carrinhoAnterior) =>
-      carrinhoAnterior.map((itemDoCarrinho) => {
-        if (itemDoCarrinho.id === novoProduto.id)
-          itemDoCarrinho.quantidade += 1;
-        return itemDoCarrinho;
-      })
-    );
-  }
   return (
     <section role="produtos" aria-label="Produtos que estão bombando!">
       <Titulo>Produtos que estão bombando!</Titulo>
